backend/requirements.txt
asgiref==3.11.0
dateutils==0.6.12
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
pillow==12.0.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
